// src/main/webapp/src/app/rest/model/rateLimit.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RateLimit { 
    limit?: number;
    remaining?: number;
    reset?: number;
    used?: number;
}